[코드 정리] 코드 가독성 개선
[내용]
- 불필요한 코드 제거
- 긴 참조 제거 및 Using 사용
- Doc ModelView 코드 순서 정리

+++ PdfToOfficeLib/PdfToOfficeLib.cpp
@@ -48,7 +48,7 @@ void PdfToOfficeLib::DoProgress(
       pProgressEventArgs->GetStatusCode();
   String statusDesc = pProgressEventArgs->GetStatusDescription();
 
-  double totalProgress = 0;
+  int totalProgress = 0;
   if (statusDesc.find(L"PDFOCRInfoText") == 0) {
     totalProgress = progress / 4;
   } else if (statusDesc.find(L"PDFLoadingInfoText") == 0) {
@@ -60,7 +60,7 @@ void PdfToOfficeLib::DoProgress(
   }
 
   if (m_ProgressSite) {
-    m_ProgressSite->SetPercent((int)totalProgress);
+    m_ProgressSite->SetPercent(totalProgress);
   }
 }
